refactor(onboarding): extract verification dialog to shared components
- Move verification dialog from UI components to dedicated folder
- Improve code organization and reusability
